+++ files/scripts/lists/brushes.lua
@@ -1,4 +1,8 @@
+dofile_once("mods/raksa/files/wands/matwand/custom_tools.lua")
+
+
 local _radial_warning="---\nWARNING: Test before applying!\n- Goes through same material\n- Can easily ruin things when not careful\n- Heavy on performance\n- But really fun to use when you get the hang of it!"
+
 
 BRUSHES = {
   {
@@ -9,6 +13,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/1_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/1_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/1_icon.png",
+    physics_supported=true,
   },
   {
     name="2px brush",
@@ -17,6 +22,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/2_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/2_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/2_icon.png",
+    physics_supported=true,
   },
   {
     name="5px brush",
@@ -25,6 +31,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/5_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/5_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/5_icon.png",
+    physics_supported=true,
   },
   {
     name="10px brush",
@@ -33,6 +40,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/10_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/10_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/10_icon.png",
+    physics_supported=true,
   },
   {
     name="20px brush",
@@ -41,6 +49,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/20_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/20_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/20_icon.png",
+    physics_supported=true,
   },
   {
     name="40px brush",
@@ -49,6 +58,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_icon.png",
+    physics_supported=true,
   },
   {
     name="80px brush",
@@ -57,6 +67,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/80_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/80_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/80_icon.png",
+    physics_supported=true,
   },
   {
     name="40px long horizontal",
@@ -65,6 +76,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_lh_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_lh_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_lh_icon.png",
+    physics_supported=true,
   },
   {
     name="40px long vertical",
@@ -73,6 +85,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_lv_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_lv_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_lv_icon.png",
+    physics_supported=true,
   },
   {
     name="40px diagonal brush 1",
@@ -81,6 +94,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_d1_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_d1_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_d1_icon.png",
+    physics_supported=true,
   },
   {
     name="40px diagonal brush 2",
@@ -89,6 +103,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_d2_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_d2_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_d2_icon.png",
+    physics_supported=true,
   },
   {
     name="40px hollow box brush",
@@ -97,6 +112,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_box_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_box_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_box_icon.png",
+    physics_supported=true,
   },
   {
     name="40px triangle-up brush",
@@ -105,6 +121,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_tri_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_tri_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_tri_icon.png",
+    physics_supported=true,
   },
   {
     name="40px triangle-right brush",
@@ -113,6 +130,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_trir_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_trir_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_trir_icon.png",
+    physics_supported=true,
   },
   {
     name="40px triangle-down brush",
@@ -121,6 +139,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_trid_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_trid_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_trid_icon.png",
+    physics_supported=true,
   },
   {
     name="40px triangle-left brush",
@@ -129,6 +148,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_tril_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_tril_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_tril_icon.png",
+    physics_supported=true,
   },
   {
     name="40px circular brush",
@@ -137,6 +157,7 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/40_cir_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/40_cir_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/40_cir_icon.png",
+    physics_supported=true,
   },
   {
     name="Cauldron",
@@ -145,6 +166,22 @@ BRUSHES = {
     reticle_file="mods/raksa/files/wands/matwand/brushes/cauldron_reticle.png",
     brush_file="mods/raksa/files/wands/matwand/brushes/cauldron_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/cauldron_icon.png",
+    physics_supported=true,
+  },
+}
+
+GROWERS = {
+  {
+    name="Tree",
+    desc=_radial_warning,
+    offset_x=5,
+    offset_y=5,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/tree_2_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/tree_2_brush.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/tree_2_icon.png",
+    click_to_use=true,
+    physics_supported=false,
+    raytrace_from_center=true,
   },
   {
     name="Small Radial Expander",
@@ -155,6 +192,8 @@ BRUSHES = {
     brush_file="mods/raksa/files/wands/matwand/brushes/radial_xs_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/radial_xs_icon.png",
     click_to_use=true,
+    physics_supported=false,
+    raytrace_from_center=true,
   },
   {
     name="Medium Radial Expander",
@@ -165,6 +204,8 @@ BRUSHES = {
     brush_file="mods/raksa/files/wands/matwand/brushes/radial_s_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/radial_s_icon.png",
     click_to_use=true,
+    physics_supported=false,
+    raytrace_from_center=true,
   },
   {
     name="Large Radial Expander",
@@ -175,6 +216,8 @@ BRUSHES = {
     brush_file="mods/raksa/files/wands/matwand/brushes/radial_m_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/radial_m_icon.png",
     click_to_use=true,
+    physics_supported=false,
+    raytrace_from_center=true,
   },
   {
     name="Huge Radial Expander",
@@ -185,7 +228,12 @@ BRUSHES = {
     brush_file="mods/raksa/files/wands/matwand/brushes/radial_l_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/radial_l_icon.png",
     click_to_use=true,
+    physics_supported=false,
+    raytrace_from_center=true,
   },
+}
+
+TOOLS = {
   {
     name="Filler Tool",
     desc="---\n[HOLD] to fill and [RELEASE] to apply\nWARNING: Can leak from edges if held too long!\nApply in small doses.",
@@ -195,26 +243,99 @@ BRUSHES = {
     brush_file="mods/raksa/files/wands/matwand/brushes/filler_brush.png",
     icon_file="mods/raksa/files/wands/matwand/brushes/filler_icon.png",
     click_to_use=true,
-    action=function(material, brush, x, y, get_draw_vars)
-      local filler = EntityCreateNew()
-
-      EntityAddComponent2(filler, "LifetimeComponent", { lifetime=2 })
-      EntityAddComponent2(
-        filler,
-        "ParticleEmitterComponent",
-        get_draw_vars("construction_paste", brush)
-      )
-
-      EntitySetTransform(filler, x, y)
-    end,
-    release_action=function(material, brush, x, y)
-      ConvertMaterialOnAreaInstantly(
-        x-1000, y-1000,
-        2000, 2000,
-        CellFactory_GetType("construction_paste"), CellFactory_GetType(material),
-        true,
-        false
-      )
-    end
+    physics_supported=true,
+    action=filler_action,
+    release_action=filler_release_action,
+  },
+  {
+    name="Line tool",
+    desc="---\n[HOLD] to draw and [RELEASE] to apply",
+    offset_x=0,
+    offset_y=0,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/0_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/scaled_brush.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/line_icon.png",
+    physics_supported=true,
+    click_to_use=false,
+    action=line_action,
+    release_action=dragger_release_action,
+  },
+  {
+    name="Rectangle tool",
+    desc="---\n[HOLD] to draw  and [RELEASE] to apply",
+    offset_x=0,
+    offset_y=0,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/0_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/scaled_brush.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/rectangle_icon.png",
+    brush_sprite_size=1,
+    physics_supported=true,
+    click_to_use=false,
+    action=corner_aligned_polygon_action,
+    release_action=dragger_release_action,
+  },
+  {
+    name="Ellipse tool",
+    desc="---\n[HOLD] to draw and [RELEASE] to apply",
+    offset_x=0,
+    offset_y=0,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/0_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/scaled_brush_circle.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/circle_icon.png",
+    brush_sprite_size=500,
+    physics_supported=true,
+    click_to_use=false,
+    action=corner_aligned_polygon_action,
+    release_action=dragger_release_action,
+  },
+  {
+    name="Rectangle tool, hollow",
+    desc="---\n[HOLD] to draw [RELEASE] to apply",
+    offset_x=0,
+    offset_y=0,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/0_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/scaled_brush_empty.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/rectangle_icon_empty.png",
+    brush_sprite_size=500,
+    physics_supported=true,
+    click_to_use=false,
+    action=corner_aligned_polygon_action,
+    release_action=dragger_release_action,
+  },
+  {
+    name="Ellipse tool, hollow",
+    desc="---\n[HOLD] to draw and [RELEASE] to apply",
+    offset_x=0,
+    offset_y=0,
+    reticle_file="mods/raksa/files/wands/matwand/brushes/0_reticle.png",
+    brush_file="mods/raksa/files/wands/matwand/brushes/scaled_brush_circle_empty.png",
+    icon_file="mods/raksa/files/wands/matwand/brushes/circle_icon_empty.png",
+    brush_sprite_size=500,
+    physics_supported=true,
+    click_to_use=false,
+    action=corner_aligned_polygon_action,
+    release_action=dragger_release_action,
   },
 }
+
+
+ALL_DRAWING_TOOLS = {
+  {
+    name="Brushes",
+    tooltip="Basic brushes for drawing",
+    brushes=BRUSHES,
+  },
+  {
+    name="Growing brushes",
+    tooltip="Time-based brushes for more exciting drawing action",
+    brushes=GROWERS,
+  },
+  {
+    name="Tools",
+    tooltip="More specialized tools",
+    brushes=TOOLS,
+  },
+}
+
+-- Helps avoid circular imports
+return ALL_DRAWING_TOOLS